move some settings and locale, setting vars and alien drop probability
alien drop probability wasn't working (var was saved but unused). Made it apply to both ore and artifacts, and renamed it to a more general name. Also changed artifact drop rate to 1/20th of alien ore drop rate. code used .9 and .045, so it's now adjusted to .9 and 1/20 * .9 (.9 passed in by function)

moved settings around a little bit for better clarity, added __ENTITY__chest-capsule-1__ to a few so they're clearer on what they affect

added "Writer" category to changelog (which doesn't load at the moment). Will be easier for Honktown (me) to show which updates I did.